requirements.txt
beautifulsoup4>=4.12.3
click>=8.1.7
cohere>=5.5.8
fastapi>=0.111.0
httpx>=0.27.0
openai>=1.35.3
openai-agents>=0.6.4
python-dotenv>=1.0.1
qdrant-client>=1.9.1
uvicorn>=0.30.1
